Updating cost kpi
